client,nginx: add support for reloading nginx
The rollback handling is incorrect at the moment: if an error occurs,
the nginx reload needs to occur *after* the nginx config files are
rolled back, not before.

Perhaps the fix is to call the rollback functions in the same order they
were added, instead of the reverse order? Or perhaps a strong guarantee
of "always reload nginx if config files changed" is required, but how?

Signed-off-by: Oliver Calder <oliver@calder.dev>