certbot,cli,client,config,proxy: add ssl configuration
Add the `--ssl` option to the `add` command, which automatically sets up
an SSL certificate using `certbot`, optionally redirecting requests from
port 80 to the listen port.

If SSL is set to redirect from port 80, and the listen port is set to
80, then setting SSL to `redirect` will automatically change the listen
port to 443.

The `ssl` setting can also be changed using the `set` command, like
other options. If the setting changes from being unset or `false` to
being `redirect` or `no-redirect`, then the `set` command calls certbot
to ensure the certificate exists and is valid.

Currently, the paths to SSL configuration and certificates are
hard-coded with respect to the listen domain. In the future, it would be
better to parse the resulting paths produced by certbot, store them in
the config, and use them when rendering the nginx configuration.

Signed-off-by: Oliver Calder <oliver@calder.dev>